requirements.txt
streamlit
pytesseract
pdf2image
python-docx
docx2txt
Pillow 